requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
azure-ai-voicelive==1.0.0b5
azure-core==1.35.1
azure-identity==1.22.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
flask==3.1.2
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
propcache==0.3.2
pycparser==2.22
pyjwt==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
yarl==1.20.1
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart
sse-starlette
jinja2
sqlalchemy==2.0.25
pymysql==1.1.0  # Driver for MySQL